Revert "More schema documentation and cleanup (#1218)"
This reverts commit f538afdbca49e34b265336aae800baa5ba685853.
