+++ schema/src/main/scala/io/shiftleft/codepropertygraph/schema/CallGraph.scala
@@ -5,7 +5,7 @@ import overflowdb.storage.ValueTypes
 
 object CallGraph extends SchemaBase {
 
-  def index: Int = 10
+  def index: Int = 6
 
   override def description: String =
     """
