+++ schema/src/main/scala/io/shiftleft/codepropertygraph/schema/ControlFlowGraph.scala
@@ -2,9 +2,9 @@ package io.shiftleft.codepropertygraph.schema
 
 import overflowdb.schema.{Cardinality, SchemaBuilder, SchemaInfo}
 
-object Cfg extends SchemaBase {
+object ControlFlowGraph extends SchemaBase {
 
-  def index: Int = 8
+  def index: Int = 6
 
   override def description: String =
     """
@@ -24,7 +24,7 @@ object Cfg extends SchemaBase {
         name = "CFG_NODE",
         comment = "Any node that can occur as part of a control flow graph"
       )
-      .addProperties(lineNumber, columnNumber)
+      .addProperties(lineNumber, columnNumber, code)
       .extendz(withinMethod, astNode)
 
     // Method and MethodReturn nodes are used as ENTRY and EXIT nodes respectively
